Add further documentation links to the README
I added these links as it took me a bit to get the project properly
configured, likely because of only a superficial knowledge of Doctrine
and its various components.

My hope is that by adding further links in the documentation, that it
will be easier and quicker for others to configure in the future.